requirements.txt
Jinja2-3.1.2 
MarkupSafe-2.1.1 
Werkzeug-2.2.2 
click-8.1.3 
colorama-0.4.6 
flask-2.2.2 
itsdangerous-2.1.2
WTForms-3.0.1 
flask-wtf-1.0.1
SQLAlchemy-1.4.45 
flask-sqlalchemy-3.0.2 
greenlet-2.0.1
Flask-Migrate-4.0.0 
Mako-1.2.4 
alembic-1.8.1